requirements.txt
pandas>=2.2
numpy>=1.26
pyyaml>=6.0.1
xgboost>=2.0
scikit-learn>=1.5
scipy>=1.11
matplotlib>=3.8
pyarrow>=15.0